# internal/docs/swagger.yaml
basePath: /api
definitions:
...
  handler.FileInfo:
    properties:
      is_dir:
        type: boolean
      mod_time:
        type: integer
      name:
        type: string
      path:
        type: string
      size:
        type: integer
    type: object
...
  handler.MkdirRequest:
    properties:
      path:
        type: string
    required:
    - path
    type: object
  handler.NewFileRequest:
    properties:
      content:
        type: string
      is_dir:
        type: boolean
      path:
        type: string
    required:
    - path
    type: object
...
  handler.RemoveRequest:
    properties:
      path:
        type: string
    required:
    - path
    type: object
...
  handler.RenameRequest:
    properties:
      new_path:
        type: string
      old_path:
        type: string
    required:
    - new_path
    - old_path
    type: object
...
  handler.TreeNode:
    properties:
      children:
        items:
          $ref: '#/definitions/handler.TreeNode'
        type: array
      is_dir:
        type: boolean
      name:
        type: string
      path:
        type: string
    type: object
...
  handler.WriteFileRequest:
    properties:
      content:
        type: string
      path:
        type: string
    required:
    - content
    - path
    type: object
host: localhost:1984
info:
  contact: {}
  description: VibeGo 后端服务 API
  title: VibeGo API
  version: 0.0.1
paths:
...
  /api/file/abs:
    get:
      parameters:
      - description: Path
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get absolute path
      tags:
      - File
  /api/file/list:
    get:
      parameters:
      - description: Directory path
        in: query
        name: path
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List directory contents
      tags:
      - File
  /api/file/mkdir:
    post:
      consumes:
      - application/json
      parameters:
      - description: Mkdir request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.MkdirRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create directory
      tags:
      - File
  /api/file/new:
    post:
      consumes:
      - application/json
      parameters:
      - description: New file request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.NewFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create new file or directory
      tags:
      - File
  /api/file/read:
    get:
      parameters:
      - description: File path
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Read file content
      tags:
      - File
  /api/file/rename:
    post:
      consumes:
      - application/json
      parameters:
      - description: Rename request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RenameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Rename file or directory
      tags:
      - File
  /api/file/rm:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Remove request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RemoveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove file or directory
      tags:
      - File
  /api/file/search:
    get:
      parameters:
      - description: Search path
        in: query
        name: path
        type: string
      - description: Search pattern
        in: query
        name: pattern
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/handler.FileInfo'
              type: array
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search files by pattern
      tags:
      - File
  /api/file/tree:
    get:
      parameters:
      - description: Directory path
        in: query
        name: path
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TreeNode'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get directory tree
      tags:
      - File
  /api/file/write:
    post:
      consumes:
      - application/json
      parameters:
      - description: Write file request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.WriteFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Write file content
      tags:
      - File
...
swagger: "2.0"